requirements.txt
asgiref==3.8.1
boto3==1.40.70
botocore==1.40.70
channels==4.0.0
dj-database-url==3.0.1
Django==5.2.7
django-admin-interface==0.30.1
django-colorfield==0.14.0
django-storages==1.14.6
et_xmlfile==2.0.0
gunicorn==23.0.0
holidays==0.86
jmespath==1.0.1
numpy==2.2.6
opencv-python-headless==4.12.0.88
openpyxl==3.1.3
packaging==25.0
pandas==2.2.2
pillow==10.3.0
psycopg2-binary==2.9.11
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0
